fix int bug

+++ discord_bot.py
@@ -123,7 +123,7 @@ async def on_message(message):
             # Send search result
             table = generate_search_result_msg(result)
             if len(table) >= 2000:
-                num_msg = ((len(table) - 1) / 2000) + 1
+                num_msg = int((len(table) - 1) / 2000) + 1
                 last_idx = 0
                 len_block = int((len(num_msg) - 1)/ num_msg) + 1
                 for _ in range(num_msg):
